deepfake/v2/requirements.txt
numpy==2.3.3
Pillow==11.3.0
timm==1.0.19
torch==2.6.0
torchvision==0.23.0
